move discord event logic from utils to discord
