+++ src/server/calender.js
@@ -1,6 +1,7 @@
 import mongoose from 'mongoose';
 import express from 'express';
-import { permission, makeDiscordEvent, deleteDiscordEvent } from './utils';
+import { permission } from './utils';
+import { deleteDiscordEvent, makeDiscordEvent } from './discord';
 import ical from 'ical-generator';
 import joi from 'joi';
 import { DateTime } from 'luxon';
